LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-pseudo-light-dismiss-invalidation-expected.txt

  option

FAIL select should not match :open when light dismissed. assert_equals: The select should match :open when opened. expected "rgb(255, 0, 0)" but got "rgb(0, 255, 0)"
